Quickload buttons added to save manager; bug fixes
Buttons for quickly loading alarms from the presets have been added to
the save manager and get saved upon preset assignment.

Improved the application startup initialization by making the default
timer the preset assigned to the first button instead of the first
alarm in the preset list.

Bug fixes:
Fixed the application crashing if the alarm audio file is missing.
Fixed timers playing the alarm when set to 00:00:00